+++ scripts/helpers.mjs
@@ -27,9 +27,13 @@ function getDistance(scene, a, b, { collisionTypes }) {
  */
 function getTokenToTokenDistance(tokenA, tokenB, { collisionTypes=[] }) {
     const scene = tokenA.parent;
-    // TODO: Gridless
-    const tokenAOffsets = tokenA.getOccupiedGridSpaceOffsets();
-    const tokenBOffsets = tokenB.getOccupiedGridSpaceOffsets();
+    // TODO: Similar lenience with gridless as gridded?
+    const tokenAOffsets = scene.grid.isGridless
+        ? [tokenA.getCenterPoint()]
+        : tokenA.getOccupiedGridSpaceOffsets();
+    const tokenBOffsets = scene.grid.isGridless
+        ? [tokenB.getCenterPoint()]
+        : tokenB.getOccupiedGridSpaceOffsets();
     // TODO: Perhaps proper elevation ranges
     const tokenAElevation = tokenA.elevation ?? 0;
     const tokenBElevation = tokenB.elevation ?? 0;
@@ -92,8 +96,54 @@ function isFinalMovementComplete(token) {
     );
 }
 
+/**
+ * Get all "source" aura effects on a given actor, split into inactive & active
+ * @param {Actor} actor                         The actor to check
+ * @returns {[ActiveEffect[], ActiveEffect[]]}  The Arrays of aura effects (active, then inactive)
+ */
+function getAllAuraEffects(actor) {
+    const activeAuras = [];
+    const inactiveAuras = [];
+    for (const effect of actor.allApplicableEffects()) {
+        if (effect.type !== "auras.aura") continue;
+        if (!effect.disabled && !effect.isSuppressed) activeAuras.push(effect);
+        else inactiveAuras.push(effect);
+    }
+    return [activeAuras, inactiveAuras];
+}
+
+/**
+ * Insert the new "aura" tab's Handlebars template part into an existing object of PARTS, for use when
+ * extending an existing AE Config Sheet
+ * @param {Record<string, HandlebarsTemplatePart} origParts     The application's original PARTS
+ * @returns {Record<string, HandlebarsTemplatePart>}            The extended PARTS
+ */
+function getExtendedParts(origParts) {
+    return Object.fromEntries(Object.entries(origParts).toSpliced(-1, 0, ["aura", { template: "modules/auras/templates/auraConfig.hbs" }]));
+}
+
+/**
+ * Add the new "aura" tab to an existing TABS object, for use when extending an existing AE Config Sheet
+ * @param {Record<string, ApplicationTabsConfiguration>} origTabs   The application's original TABS
+ * @returns {Record <string, ApplicationTabsConfiguration>}         The extended TABS
+ */
+function getExtendedTabs(origTabs) {
+    return {
+        sheet: {
+            ...origTabs.sheet,
+            tabs: [
+                ...origTabs.sheet.tabs,
+                { id: "aura", icon: "fa-solid fa-person-rays" }
+            ]
+        }
+    };
+}
+
 export {
     getNearbyTokens,
     getTokenToTokenDistance,
-    isFinalMovementComplete
+    isFinalMovementComplete,
+    getAllAuraEffects,
+    getExtendedParts,
+    getExtendedTabs
 };