+++ scripts/plugins/pluginHelpers.mjs
@@ -1,0 +1,9 @@
+import replaceDAESheet from "./dae.mjs";
+
+function overrideSheets() {
+    replaceDAESheet();
+}
+
+export {
+    overrideSheets
+};