+++ scripts/settings.mjs
@@ -7,4 +7,12 @@ export function registerSettings() {
         type: Boolean,
         default: false
     });
+    game.settings.register("auras", "exactCircles", {
+        name: "AURAS.SETTINGS.ExactCircles.Name",
+        hint: "AURAS.SETTINGS.ExactCircles.Hint",
+        scope: "client",
+        config: true,
+        type: Boolean,
+        default: false
+    });
 }