+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Auras Changelog
 
+## Version 0.3.0
+- Added the beginning of plugin behavior (so far just compatibility with DAE's custom sheet)
+- Setting for showing circles on gridded maps with "Exact" set as diagonal distances
+- Settings for whether an aura should apply to self or not
+- Added logic for the above - depends on a V13 issue being addressed
+- Added logic for effects being enabled/disabled, deleted, and to ensure old effects (from no-longer-active/existent auras) are properly wiped on token movement
+
 ## Version 0.2.0
 - Added aura visualization
 - Added changelog